requirements.txt
Django==3.1.4
django-allauth==0.44.0
django-ckeditor==6.0.0
django-crispy-forms==1.10.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-inlinecss==0.3.0
django-cors-headers==3.6.0
djangorestframework==3.12.2
django-filter==2.4.0
gunicorn==20.0.4
Pillow==5.1.0
python-dotenv==0.15.0
django-tinymce==3.2.0
whitenoise==5.2.0
